Split staking constants and helpers into its own submodule
